+++ plugin-dev/Source/Sentry/Private/SentrySubsystem.cpp
@@ -15,6 +15,7 @@
 #include "Misc/EngineVersion.h"
 #include "Misc/CoreDelegates.h"
 #include "Misc/App.h"
+#include "GenericPlatform/GenericPlatformDriver.h"
 #include "GenericPlatform/GenericPlatformMisc.h"
 
 #include "Interface/SentrySubsystemInterface.h"
@@ -89,6 +90,12 @@ void USentrySubsystem::Initialize()
 	}
 
 	AddDefaultContext();
+
+#if PLATFORM_WINDOWS || PLATFORM_LINUX || PLATFORM_MAC
+	AddGpuContext();
+	AddDeviceContext();
+#endif
+
 	PromoteTags();
 	ConfigureBreadcrumbs();
 }
@@ -286,6 +293,31 @@ void USentrySubsystem::AddDefaultContext()
 	DefaultContext.Add(TEXT("Game name"), FApp::GetName());
 
 	SubsystemNativeImpl->SetContext(TEXT("Unreal Engine"), DefaultContext);
+}
+
+void USentrySubsystem::AddGpuContext()
+{
+	FGPUDriverInfo GpuDriverInfo = FPlatformMisc::GetGPUDriverInfo(FPlatformMisc::GetPrimaryGPUBrand());
+
+	TMap<FString, FString> GpuContext;
+	GpuContext.Add(TEXT("name"), GpuDriverInfo.DeviceDescription);
+	GpuContext.Add(TEXT("vendor_name"), GpuDriverInfo.ProviderName);
+	GpuContext.Add(TEXT("driver_version"), GpuDriverInfo.UserDriverVersion);
+
+	SubsystemNativeImpl->SetContext(TEXT("gpu"), GpuContext);
+}
+
+void USentrySubsystem::AddDeviceContext()
+{
+	const FPlatformMemoryConstants& MemoryConstants = FPlatformMemory::GetConstants();
+
+	TMap<FString, FString> DeviceContext;
+	DeviceContext.Add(TEXT("cpu_description"), FPlatformMisc::GetCPUBrand());
+	DeviceContext.Add(TEXT("number_of_cores"), FString::FromInt(FPlatformMisc::NumberOfCores()));
+	DeviceContext.Add(TEXT("number_of_cores_including_hyperthreads"), FString::FromInt(FPlatformMisc::NumberOfCoresIncludingHyperthreads()));
+	DeviceContext.Add(TEXT("physical_memory_size_gb"), FString::FromInt(MemoryConstants.TotalPhysicalGB));
+
+	SubsystemNativeImpl->SetContext(TEXT("device"), DeviceContext);
 }
 
 void USentrySubsystem::PromoteTags()