+++ plugin-dev/Source/Sentry/Public/SentrySubsystem.h
@@ -234,6 +234,12 @@ private:
 	/** Adds default context data for all events captured by Sentry SDK. */
 	void AddDefaultContext();
 
+	/** Adds GPU context data for all events captured by Sentry SDK. */
+	void AddGpuContext();
+
+	/** Adds GPU context data for all events captured by Sentry SDK. */
+	void AddDeviceContext();
+
 	/** Promote specified values to tags for all events captured by Sentry SDK. */
 	void PromoteTags();