+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add scope support for Windows/Linux ([#328](https://github.com/getsentry/sentry-unreal/pull/328))
 - Add extra crash context for native integration ([#342](https://github.com/getsentry/sentry-unreal/pull/342))
 - Add missing plugin settings ([#335](https://github.com/getsentry/sentry-unreal/pull/335))
+- Update event context categories for desktop ([#356](https://github.com/getsentry/sentry-unreal/pull/356))
 
 ### Fixes